campaign_generator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "campaign-generator"
version = "0.1.0"
description = "Generate pre-authored TTRPG campaigns from a genre pack and campaign seed."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "python-dotenv>=1.0,<2.0",
  "pyyaml>=6.0,<7.0",
  "rich>=13.7,<14.0",
  "tenacity>=9.0,<10.0",
  "typer>=0.12,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
campaign-generator = "campaign_generator.__main__:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["campaign_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
